IR building

+++ src/MxCompiler/IR/IRnodes/intImd.java
@@ -3,13 +3,13 @@ package MxCompiler.IR.IRnodes;
 import java.math.BigInteger;
 
 public class intImd extends intValue{
-    private BigInteger value;
+    private int value;
 
-    public intImd(BigInteger value){
+    public intImd(int value){
         this.value = value;
     }
 
-    public BigInteger getValue() {
+    public int getValue() {
         return value;
     }
 }
